pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ph-engagement-bot"
version = "1.0.0"
description = "Semi-automated Product Hunt engagement with Telegram approval"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your@email.com"}
]
keywords = ["product-hunt", "telegram", "automation", "engagement"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "python-telegram-bot>=20.0",
    "anthropic>=0.18.0",
    "apscheduler>=3.10.0",
    "httpx>=0.25.0",
    "beautifulsoup4>=4.12.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[project.scripts]
ph-engagement = "ph_engagement.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]
